1) bugfix with missing NGPUS; 2) propagated the change for subdirectory into all mpiexec scripts.

+++ mpiexec-impixtmp.sh
@@ -28,39 +28,41 @@ else
     #LD_LIBRARY_PATH=`echo $LD_LIBRARY_PATH | sed 's!/opt/cuda[^:]*:!:!g'`
 
     echo ">>> PSUBMIT: mpiexec is: " $(which mpiexec)
-    echo ">>> PSUBMIT: Executable is: " $(which $PSUBMIT_SUBDIR/$TARGET_BIN)
-#    echo ">>> PSUBMIT: ldd:"
-#    ldd $(which $PSUBMIT_SUBDIR/$TARGET_BIN)
     export PSUBMIT_JOBID PSUBMIT_NP
     [ -z "$PSUBMIT_PREPROC" ] || eval $PSUBMIT_PREPROC
 
-    [ -f "hostfile.$PSUBMIT_JOBID" ] && machinefile="-machinefile hostfile.$PSUBMIT_JOBID"
+    if [ "$TARGET_BIN" != "false" ]; then
+        [ -f "hostfile.$PSUBMIT_JOBID" ] && machinefile="-machinefile hostfile.$PSUBMIT_JOBID"
+        executable=$PSUBMIT_SUBDIR/$TARGET_BIN
+        if [ ! -e $executable ]; then
+            executable=$(which $TARGET_BIN)
+        fi
+        echo ">>> PSUBMIT: Executable is: " $executable
+        if [ ! -z "$executable" ]; then
+            newexecname=$(mktemp)
+            nds=$(echo $PSUBMIT_NODELIST | sed 's/,/ /g')
+            for i in $nds; do 
+                node=$(echo $i | cut -d: -f1); 
+                scp "$executable" ${node}:$newexecname; 
+                ssh ${node} "chmod +x $newexecname"; 
+            done
 
-	[ ! -f $PSUBMIT_SUBDIR/$TARGET_BIN ] && TARGET_BIN=$(which $TARGET_BIN)
+            time2=$(date +"%s");
+            export I_MPI_HYDRA_BOOTSTRAP="ssh"
+            echo $- | grep -q x && omit_setx=true || set -x
+            mpiexec.hydra $machinefile -np "$PSUBMIT_NP" -ppn "$PSUBMIT_PPN" --errfile-pattern=err.$PSUBMIT_JOBID.%r --outfile-pattern=out.$PSUBMIT_JOBID.%r "$newexecname" $ALL_ARGS
+            { [ -z "$omit_setx" ] && set +x; } 2>/dev/null
 
-	newexecname=$(mktemp)
-	nds=$(echo $PSUBMIT_NODELIST | sed 's/,/ /g')
-	for i in $nds; do 
-		node=$(echo $i | cut -d: -f1); 
-		scp "$TARGET_BIN" ${node}:$newexecname; 
-		ssh ${node} "chmod +x $newexecname"; 
-	done
+            for i in $nds; do 
+                node=$(echo $i | cut -d: -f1); 
+                ssh $node "rm -f $newexecname"; 
+            done
 
-    time2=$(date +"%s");
-    export I_MPI_HYDRA_BOOTSTRAP="ssh"
-    echo $- | grep -q x && omit_setx=true || set -x
-	mpiexec.hydra $machinefile -np "$PSUBMIT_NP" -ppn "$PSUBMIT_PPN" --errfile-pattern=err.$PSUBMIT_JOBID.%r --outfile-pattern=out.$PSUBMIT_JOBID.%r "$newexecname" $ALL_ARGS
-    [ -z "$omit_setx" ] && set +x
-
-	for i in $nds; do 
-		node=$(echo $i | cut -d: -f1); 
-		ssh $node "rm -f $newexecname"; 
-	done
-
-    time3=$(date +"%s");
-    walltime="$(expr $time3 - $time2)"
-    [ "$(expr $time3 - $time1)" -lt "2" ] && sleep $(expr 2 - $time3 + $time1)
-    echo ">>> PSUBMIT: Walltime: $walltime"
-
+            time3=$(date +"%s");
+            walltime="$(expr $time3 - $time2)"
+            [ "$(expr $time3 - $time1)" -lt "2" ] && sleep $(expr 2 - $time3 + $time1)
+            echo ">>> PSUBMIT: Walltime: $walltime"
+        fi   
+    fi
     [ -z "$PSUBMIT_POSTPROC" ] || eval $PSUBMIT_POSTPROC
 fi
